feat(result): introduce max results flag
This commit introduces a new flag, `max`, to the `Scrape` function in the `urban-cli` library. This parameter allows users to specify the maximum number of results to be returned by the function. This is useful for users who only want to see a limited number of results or who want to avoid the performance overhead of fetching too many results.

The `max` parameter is an integer value that defaults to 7. If the value of the `max` parameter is greater than 0, the `Scrape` function will only return up to `max` results. If the value of the `max` parameter is 0 or less, the `Scrape` function will return all results.

In addition to introducing the `max` parameter, this commit also makes some minor changes to the `Scrape` function:

- The function now uses a `colly.Collector` with a `MaxDepth` value of `(max-1)/7 + 1`. This ensures that the function will not crawl more than `max` pages.
- The function now adds a "Next page" button to the HT